--- packages/foundry-react-ui/src/enums/checkboxTypes.ts
@@ -1,9 +1,0 @@
-export enum CheckboxTypes {
-    fill = 'fill',
-    cross = 'cross',
-    check = 'check',
-    default = 'default',
-    neutral = 'neutral',
-}
-
-export default CheckboxTypes;
