+++ packages/foundry-react-ui/src/index.ts
@@ -19,7 +19,6 @@ import timings from './enums/timings';
 import fonts from './enums/fonts';
 import variants from './enums/variants';
 import feedbackTypes from './enums/feedbackTypes';
-import checkboxTypes from './enums/checkboxTypes';
 
 export {
   Button,
@@ -44,5 +43,4 @@ export {
   fonts,
   variants,
   feedbackTypes,
-  checkboxTypes
 };
